refactor: cria controller de usuário
