--- src/teste5.ts
@@ -1,9 +1,0 @@
-
-
-module.exports = function(req, res){
-    
-    var name =  req.query.name;
-
-    res.send("Usuário " +  name  + "  foi lido 0 vezes.");
-
-};
